tools/maintainers/requirements.txt
PyYAML>=6.0.1,<7.0.0
requests>=2.31.0,<3.0.0
